OnnixBot with Payments/crypto-payment-bot/requirements.txt
# Crypto Payment Bot Requirements
python-telegram-bot==22.5
yfinance==0.2.66
pandas>=2.2.0
numpy>=1.24.0
python-dotenv>=1.2.0
requests>=2.31.0
cryptography>=41.0.0
qrcode>=7.4.0
flask>=2.3.0
Pillow>=10.0.0